add redundancy to jamaicans under rule 3.
This adds another mention of jamaicans to rule 3 directly, because people were apparently missing it's mention in an un-expected place within the rules. 

I also fixed formatting on the line 'The use of the word "retard" and its derivatives is not allowed. The only exception to this rule is the term "deltard" when used in an appropriate contextual setting.'